미션 도전하기

+++ hajin/src/main/java/umc/study/repository/MissionRepository/MissionRepository.java
@@ -3,6 +3,6 @@ package umc.study.repository.MissionRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.study.domain.Mission;
 
-public interface MissionRepository extends JpaRepository<Mission, Integer>, MissionRepositoryCustom{
+public interface MissionRepository extends JpaRepository<Mission, Long>, MissionRepositoryCustom{
 
 }
